pyproject.toml
[project]
name = "carf"
version = "0.1.0"
description = "Complex-Adaptive Reasoning Fabric: A Neuro-Symbolic-Causal Agentic System"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "CARF Team"}
]
keywords = ["agentic-ai", "causal-inference", "langgraph", "neuro-symbolic"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Core Orchestration
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",

    # Type Safety & Validation
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",

    # Reliability & Resilience
    "tenacity>=8.2.0",

    # Human-in-the-Loop
    "humanlayer>=0.6.0",

    # API Framework (MVP)
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",

    # Configuration
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",

    # Graph Database (Phase 3)
    "neo4j>=5.15.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]
causal = [
    # Causal Inference (Phase 2)
    "dowhy>=0.11.0",
    "econml>=0.15.0",
    "causal-learn>=0.1.3",
]
bayesian = [
    # Bayesian Inference (Phase 2)
    "pymc>=5.10.0",
    "arviz>=0.17.0",
]
dashboard = [
    # Streamlit cockpit
    "streamlit>=1.32.0",
]
kafka = [
    # Audit trail
    "confluent-kafka>=2.3.0",
]
router = [
    # Router model training
    "accelerate>=0.27.0",
    "datasets>=2.19.0",
    "scikit-learn>=1.4.0",
    "torch>=2.1.0",
    "transformers>=4.41.0",
]
databases = [
    # Database Integrations
    "neo4j>=5.15.0",
    "redis>=5.0.0",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
]
observability = [
    # Tracing & Monitoring
    "langsmith>=0.1.0",
    "opentelemetry-api>=1.22.0",
    "opentelemetry-sdk>=1.22.0",
]
all = [
    "carf[dev,causal,bayesian,dashboard,kafka,router,databases,observability]",
]

[project.scripts]
carf = "src.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = ["E501"]  # Line too long (handled by formatter)

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=src --cov-report=term-missing"
